RANCHER-1950. deployModuleFromFeatureBranch was refactored, issue affected by recent PR was fixed. (#731)
* RANCHER-1950. Implemented the addTenant method in EurekaNamespace to support application updating.

* RANCHER-1950. The method getExistedTenants has been improved to better identify tenants with specific modules. The code for updating the application and its modules has also been improved

* RANCHER-1950. Add method to retrieve deployment container image.

Introduce two new functions: `getDeploymentContainerImage` and `getDeploymentContainerImageName` in `kubectl.groovy`. These functions retrieve the full image path of a container within a deployment and extract just the image name, respectively. The changes enhance the script's capability to handle deployment configurations efficiently.

* RANCHER-1950. Add stage to retrieve module's sidecar image

A new stage has been introduced to the Jenkins pipeline to fetch the sidecar image for a module. This is done using the kubeconfig to query and retrieve the sidecar image, which is then added to the namespace's module list.

* RANCHER-1950. Format sidecarImage correctly in Jenkinsfile

Ensure that the sidecarImage uses hyphens instead of colons by replacing ':' with '-' in its name. This change is necessary to maintain compatibility with systems that do not handle colons well. Additionally, a debug log statement was added to output the namespace modules configuration.

* RANCHER-1950. Remove an old approach regarding the Modules class and class allModules property.

* RANCHER-1950. Minimize direct EurekaModule parameter's updates. The common approach for that is to set up them via the loadModuleDetails method.

* RANCHER-1950. Add updated module to the tenant module list.